Update permissions to only be accessible to super-admins
